+++ DESIRE/basemetric.h
@@ -7,9 +7,9 @@
 #define __Basemetric
 
 
-float baseme(float* a1, float* a2, char* s1, char* s2, int m,int size);
+float baseme(float* a1, float* a2, char* s1, char* s2, long long m,long long size);
 float trans(float ddd1, float ddd2);
 float trans2(float ddd1, float ddd2);
-extern int* metricm;
+extern long long* metricm;
 extern float* metricmaxdis;
 #endif
