+++ DESIRE/bptree.h
@@ -12,48 +12,48 @@
 #include<queue>
 #include<cstdlib>
 #include "iopack/RAF.h"
-#include "iopack\blk_file.h"
+#include "iopack/blk_file.h"
 #include "object.h"
 #include "bpnode.h"
 using namespace std;
-extern int f;
+extern long long f;
 extern float eps;
 class bptree //: public Cacheable
 {
 public:
 	/*
-	
-	
 
-	
-	
-	
+
+
+
+
+
 	void delnode();
 	void update();
-	
+
 	*/
 	//bptree* querydis(float d);
-	int block, dirty, fa, isleaf, leafnum;
+	long long block, dirty, fa, isleaf, leafnum;
 	float dis;
-	float tmpdis;	
+	float tmpdis;
 	bpnode** son;
 	void outnode(string str);
-	int getsize();
-	int write_to_buffer(char* buffer);
-	int read_from_buffer(char* buffer);
+	long long getsize();
+	long long write_to_buffer(char* buffer);
+	long long read_from_buffer(char* buffer);
 	void flush();
 	void bfs(string str);
-	void loadroot(int page);	
+	void loadroot(long long page);
 	void insert(bpnode* q);
-	bptree* prenode(int loc);
+	bptree* prenode(long long loc);
 	bptree* lastnode();
-	void delnode(int loc);
+	void delnode(long long loc);
 	void update();
 	bptree();
 	~bptree();
 
 	void entere(bpnode* q);
-	void split(int flag,bptree* rt);
+	void split(long long flag,bptree* rt);
 
 };
