+++ DESIRE/iopack/RAF.h
@@ -8,13 +8,13 @@ class Object;
 class RAF: public Cacheable
 {
 public:
-	int num_obj; // the total number of objects
-	int last_ptr;
-	void init(char *_fname, int _b_length, Cache *_c);
+	long long num_obj; // the total number of objects
+	long long last_ptr;
+	void init(char *_fname, long long _b_length, Cache *_c);
 	void init_restore(char *_fname, Cache *_c);
-	int add_object(Object* obj);
-	int* buid_from_array(Object** objset);
-	Object* get_object(int ptr);
+	long long add_object(Object* obj);
+	long long* buid_from_array(Object** objset);
+	Object* get_object(long long ptr);
 };
 
 #endif
